third_party/WebKit/LayoutTests/external/wpt/fetch/api/request/request-disturbed-expected.txt

This is a testharness.js-based test.
FAIL Request's body: initial state assert_equals: body's default value is null expected (object) null but got (undefined) undefined
PASS Request without body cannot be disturbed
PASS Check cloning a disturbed request
PASS Check creating a new request from a disturbed request
FAIL Check creating a new request with a new body from a disturbed request Failed to construct 'Request': Cannot construct a Request with a Request object that has already been used.
FAIL Input request used for creating new request became disturbed assert_not_equals: body should not be undefined got disallowed value undefined
FAIL Input request used for creating new request became disturbed even if body is not used assert_not_equals: body should not be undefined got disallowed value undefined
PASS Check consuming a disturbed request
PASS Request construction failure should not set "bodyUsed"
Harness: the test ran to completion.
